Fix recursive when URL contains explicit port
Previously when handling recursive jobs, found folders were
not being added if the user supplied url (-u) contained a
port number (ie, http://example.com:80/FUZZ).

This is because the logic to determine if a redirect was a
folder was to compare if the redirected URL == the
request-url + "/".  (ie, is a trailing "/" added)

The issue is that GetRedirectLocation() strips the port,
which means the if statement is wrongly always false (as the
port is missing).

This change now checks if the redirection has a trailing
"/", which means we can imply the redirection is a folder,
as that is the real factor that needs testing.

The issue with this fix is that if a redirect goes to a
whole new url, then we have *potentially* implemented follow
redirects on recursive (ffuf/ffuf#192) which is probably not
desirable?

Fixes ffuf/ffuf#232

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>