More dynamic solutions
ffuf now handles the progress dynamically. if a string is too long for the width of the terminal it will switch over to a minified version. This can also work vice versa if the string because shorter than the width it will full back to the orginal full progress string.

If the minified version is still too short of the terminal ffuf will now flick over to Quite mode not printing any progress.