introduces the config package to refactor the way the config is obtained
This commit refactors the ffuf code to simplify the configuration
process. ffuf's configuration has become complex, since it takes configurations
from multiple configuration sources like the commandline, multiple configuration
files and HTTP request files (sometimes called HTTP templates). Inputs from
these configuration sources have to be merged into one coherent configuration
and it's consistency validated. ffuf lacked a clear way to obtain it's
configuration easily and clearly. Any contributor who wanted to introduce a new
feature had to study the configuration code and insert his or modify the
configuration process with the hopes of not breaking anything.

The config package aims to simplify and formalize the configuration of ffuf,
while making it flexible and extensible. It provides:
  * One simple function call to obtain a configuration from a standard hierarchy
    of all configurations sources
  * A mechanism to merge the defined configuration sources
  * A mechanism to validate the merged configuration and supply error messages
    to the caller
  * A way to add a configuration source to ffuf
  * A way to add validation to new or changed options without knowing the code
    base

A pkg/config/README.md is provided to further explain the configuration
process. Additionally, the core functionality is of this package is tested an
examples as Go code are provided.

This commit had to make changes to code outside the main package to untangle the
imports, which were causing import cycles.