adjusts request file parsing: fixes bugs and changes version parsing
The HTTP version was parsed with http.ParseHTTPVersion, which conforms to RFC
7230, which makes explicit, that the version part of the request line must have
a major and minor version like so "HTTP/<major>.<minor>", http.ParseHTTPVersion
errors if the version part is "HTTP/2".

Because this version of ffuf extracts configuration parameters from the version
part in the request line, especially ConfigOptions.HTTPOptions.Http2, a helper
function parseHTTPVersion was introduced, which handles the version extraction
more suitable to web fuzzing.